feat: implement modules/cjsToEsm
